iot-app/dlClassifier/requirements.txt
scikit_learn==0.24.1
matplotlib
numpy
keras
tensorflow
redis
pyinotify
python-dotenv